requirements.txt
flask
flask-cors
requests
gunicorn
google-generativeai